// src/interface.ts
export enum LiveRoomTypeEnum {
  /** 系统推流 */
  system = 1,
  /** 主播使用srs推流 */
  srs,
  /** 主播使用obs/ffmpeg推流 */
  obs,
  /** 主播使用webrtc推流，直播 */
  wertc_live,
  /** 主播使用webrtc推流，会议，实现一 */
  wertc_meeting_one,
  /** 主播使用webrtc推流，会议，实现二 */
  wertc_meeting_two,
  /** 主播使用msr推流 */
  msr,
  /** 主播打pk */
  pk,
  /** 主播使用腾讯云css推流 */
  tencentcloud_css,
  /** 主播使用腾讯云css推流打pk */
  tencentcloud_css_pk,
  /** 转推b站 */
  forward_bilibili,
  /** 转推虎牙 */
  forward_huya,
  /** 转推斗鱼 */
  forward_douyu,
  /** 转推抖音 */
  forward_douyin,
  /** 转推抖音 */
  forward_kuaishou,
  /** 转推小红书 */
  forward_xiaohongshu,
  /** 转推所有 */
  forward_all,
}
// ...
export enum LiveRoomStatusEnum {
  /** 正常 */
  normal,
  /** 禁用 */
  disable,
}

/** 直播间是否显示 */
export enum LiveRoomIsShowEnum {
  /** 显示 */
  yes,
  /** 不显示 */
  no,
}

/** 直播间是否开启聊天 */
export enum LiveRoomOpenChatEnum {
  yes,
  no,
}
// ...
export enum LiveRoomTipTouristLoginEnum {
  /** 关闭，即不提醒游客登录 */
  close,
  /** 开启提醒游客登录，游客可关闭 */
  open,
  /** 强制游客登录，即游客不可关闭 */
  force,
}

export interface ILiveRoom {
  id?: number;
  /** 标题 */
  title?: string;
  /** 简介 */
  desc?: string;
  /** 权重 */
  priority?: number;
  /** 推流秘钥 */
  key?: string;
  /** 类型 */
  type?: LiveRoomTypeEnum;
  /** 封面图 */
  cover_img?: string;
  /** 关键帧图 */
  keyframe_img?: string;
  /** 背景图 */
  bg_img?: string;
  /** 状态 */
  status?: LiveRoomStatusEnum;
  /** 是否显示 */
  is_show?: SwitchEnum;
  /** 是否开启聊天 */
  open_chat?: SwitchEnum;
  /** 提醒游客登录 */
  tip_tourist_login?: LiveRoomTipTouristLoginEnum;
  /** 提醒游客登录间隔 */
  tip_tourist_login_delay?: number;
  /** 游客能否发言 */
  tourist_send_msg?: SwitchEnum;
  /** 聊天关键词过滤 */
  keyword_filter_msg?: string;
  /** 用户名关键词过滤 */
  keyword_filter_username?: string;
  /** 历史消息条数 */
  history_msg_total?: number;
  /** 新注册用户不能发言时间(秒) */
  newuser_send_msg_delay?: number;
  /** 房间密码 */
  room_pwd?: string;
  /** 关闭房间 */
  is_close?: SwitchEnum;
  /** 关闭房间描述 */
  is_close_desc?: string;
  /** 发送消息间隔(秒) */
  send_msg_throttle?: number;
  /** 是否开启签到 */
  is_show_signin?: SwitchEnum;
  /** 是否开启手机看直播 */
  is_show_phone_live?: SwitchEnum;
  /** 公告通知 */
  notice_msg?: string;
  /** 系统消息 */
  system_msg?: string;
  /** 显示在线人数 */
  is_show_live_user_nums?: SwitchEnum;
  /** 设置最低在线人数 */
  mock_live_user_nums_min?: number;
  /** 设置最高在线人数 */
  mock_live_user_nums_max?: number;
  /** 最在线人数刷新间隔 */
  mock_live_user_nums_refresh_delay?: number;
  /** 聊天消息审核 */
  msg_verify?: SwitchEnum;
  is_fake?: SwitchEnum;

  pull_rtmp_url?: string;
  pull_flv_url?: string;
  pull_hls_url?: string;
  pull_webrtc_url?: string;

  pull_cdn_rtmp_url?: string;
  pull_cdn_flv_url?: string;
  pull_cdn_hls_url?: string;
  pull_cdn_webrtc_url?: string;

  push_rtmp_url?: string;
  push_obs_server?: string;
  push_obs_stream_key?: string;
  push_webrtc_url?: string;
  push_srt_url?: string;

  push_cdn_rtmp_url?: string;
  push_cdn_obs_server?: string;
  push_cdn_obs_stream_key?: string;
  push_cdn_webrtc_url?: string;
  push_cdn_srt_url?: string;

  forward_bilibili_url?: string;
  forward_douyu_url?: string;
  forward_huya_url?: string;
  forward_douyin_url?: string;
  forward_kuaishou_url?: string;
  forward_xiaohongshu_url?: string;

  /** 直播间备注 */
  remark?: string;

  areaId?: number;

  /** 用户信息 */
  user?: IUser;
  /** 用户信息 */
  users?: IUser[];
  /** 分区信息 */
  area?: IArea;
  /** 分区信息 */
  areas?: IArea[];
  /** 直播信息 */
  live?: ILive;
  user_live_room?: IUserLiveRoom & { user: IUser };

  created_at?: string;
  updated_at?: string;
  deleted_at?: string;
}
// ...
export interface IQqUser {
  id?: number;
  client_id?: number;
  openid?: string;
  unionid?: string;
  nickname?: string;
  figureurl?: string;
  figureurl_1?: string;
  figureurl_2?: string;
  figureurl_qq_1?: string;
  figureurl_qq_2?: string;
  constellation?: string;
  gender?: string;
  city?: string;
  province?: string;
  year?: string;
  ret?: number;
  created_at?: string;
  updated_at?: string;
  deleted_at?: string;
}

export interface IWechatUser {
  id?: number;
  appid?: string;
  openid?: string;
  nickname?: string;
  sex?: number;
  province?: string;
  city?: string;
  country?: string;
  headimgurl?: string;
  privilege?: string;
  unionid?: string;
  created_at?: string;
  updated_at?: string;
  deleted_at?: string;
}

export interface IEmailUser {
  id?: number;
  email?: string;
  created_at?: string;
  updated_at?: string;
  deleted_at?: any;
}

export interface IGithubUser {
  id?: number;
  client_id?: string;
  login?: string;
  github_id?: number;
  node_id?: string;
  avatar_url?: string;
  gravatar_id?: string;
  url?: string;
  html_url?: string;
  type?: string;
  site_admin?: string;
  name?: string;
  company?: string;
  blog?: string;
  location?: string;
  email?: any;
  hireable?: any;
  bio?: string;
  twitter_username?: any;
  public_repos?: number;
  public_gists?: number;
  followers?: number;
  following?: number;
  github_created_at?: string;
  github_updated_at?: string;
  private_gists?: number;
  total_private_repos?: number;
  owned_private_repos?: number;
  disk_usage?: number;
  collaborators?: number;
  two_factor_authentication?: string;
  created_at?: string;
  updated_at?: string;
  deleted_at?: any;
}

export interface IThirdUser {
  id?: number;
  user_id?: number;
  third_user_id?: number;
  third_platform?: number;
  created_at?: string;
  updated_at?: string;
  deleted_at?: string;
}

/** 用户状态 */
export enum UserStatusEnum {
  /** 正常 */
  normal,
  /** 禁用 */
  disable,
}

export enum QuizMatchStatus {
  notStarted = 0, // 未开始
  inProgress = 1, // 进行中
  completed = 2, // 已结束
}

export enum UserIsTouristEnum {
  yes,
  no,
}

export interface IUser {
  id?: number;
  username?: string;
  password?: string;
  status?: UserStatusEnum;
  avatar?: string;
  desc?: string;
  token?: string;
  agent_account_for?: number;
  is_tourist?: UserIsTouristEnum;

  qq_users?: IQqUser[];
  wechat_users?: IWechatUser[];
  github_users?: IGithubUser[];
  email_users?: IEmailUser[];

  roles?: any[];
  auths?: any[];
  is_admin?: SwitchEnum;
  is_agent?: SwitchEnum;

  user_roles?: number[];
  wallet?: IWallet;

  live_room?: ILiveRoom;
  live_rooms?: ILiveRoom[];

  created_at?: string;
  updated_at?: string;
  deleted_at?: string;
}
// ...
export enum FormTypeEnum {
  'input' = 'input',
  'password' = 'password',
  'number' = 'number',
  'select' = 'select',
  'radio' = 'radio',
  'checkbox' = 'checkbox',
  'markdown' = 'markdown',
  'switch' = 'switch',
  'upload' = 'upload',
  'treeSelect' = 'treeSelect',
  'datePickerRange' = 'datePickerRange',
  'datePickerDatetime' = 'datePickerDatetime',
}
// ...
export interface IWallet {
  id?: number;
  user_id?: number;
  balance?: string;
  created_at?: string;
  updated_at?: string;
  deleted_at?: string;
}
// ...
export interface IAreaLiveRoom {
  id?: number;
  area_id?: number;
  live_room_id?: number;
  /** 分区信息 */
  area?: IUser;
  /** 直播间信息 */
  live_room?: ILiveRoom;
  created_at?: string;
  updated_at?: string;
  deleted_at?: string;
}

export interface IArea {
  id?: number;
  name?: string;
  /** 备注 */
  remark?: string;
  /** 权重 */
  priority?: number;
  area_live_rooms?: IAreaLiveRoom[];
  live_room_is_show?: LiveRoomIsShowEnum;
  live_room_status?: LiveRoomStatusEnum;
  created_at?: string;
  updated_at?: string;
  deleted_at?: string;
}

export interface ISrsPublishStream {
  /** 客户端在获取信息时，必须检查ServerID是否改变，改变时就是服务器重启，之前所有的数据都应该作废了。 */
  srs_server_id?: string;
  srs_service_id?: string;
  srs_action?: string;
  srs_client_id?: string;
  srs_ip?: string;
  srs_vhost?: string;
  srs_app?: string;
  srs_tcUrl?: string;
  srs_stream?: string;
  srs_param?: string;
  srs_stream_url?: string;
  srs_stream_id?: string;
}

export type ILive = {
  id?: number;
  /** 直播平台 */
  platform?: LivePlatformEnum;
  /** 直播流名称 */
  stream_name?: string;
  /** 直播流id */
  stream_id?: string;
  /** 用户id */
  user_id?: number;
  /** 直播间id */
  live_room_id?: number;
  /** 备注 */
  remark?: string;

  /** 直播间信息 */
  live_room?: ILiveRoom;
  /** 用户信息 */
  user?: IUser;

  created_at?: string;
  updated_at?: string;
  deleted_at?: string;
};

/** 直播平台 */
export enum LivePlatformEnum {
  rtc,
  srs,
  tencentcloud_css,
}
// ...
export interface IUserLiveRoom {
  id?: number;
  user_id?: number;
  live_room_id?: number;
  /** 用户信息 */
  user?: IUser;
  /** 直播间信息 */
  live_room?: ILiveRoom;
  created_at?: string;
  updated_at?: string;
  deleted_at?: string;
}
// ...
export enum SwitchEnum {
  yes,
  no,
}